Update patch set 14
Patch Set 14:

I verified binary that a backup/restore/backup contents are binary equivalent 

Chenglu did the same. He also informed me it doesn't leak after many backup/restore operations + he verified selinux permissions are maintained

Only leading / is present but i suspect it is part of tar standard until we add '-C /' option. It is same in twrp backups since they moved to libtar. If it is not how it does in gnu tar, we can alter it later in a separate commit as it doesn't impact in recovery 

For me it is ready for merging + the recovery part

Patch-set: 14