Update patch set 14
Patch Set 14:

Hope I am not posting some crap again:

From GNU tar:

"The magic field indicates that this archive was output in the P1003 archive format. If this field contains TMAGIC, the uname and gname fields will contain the ASCII representation of the owner and group of the file respectively. If found, the user and group IDs are used rather than the values in the uid and gid fields."

So, if I understand it right, by default, a gnu tar will prefer name over numerical id

In that case, a command option is in fact preferred so that libtar stays usable outside recovery with a reliable gnu tar like behavior

Patch-set: 14