Update patch set 1
Patch Set 1:

Sorry I'm not following your logic.  This change is for extract, and it seems like you are concerned about create/append...?

Note this change unlinks the new file (if it exists) before attempting link().  It does not unlink the existing file.  See "man 2 link": "If newpath exists it will not be overwritten."

This only matters when extracting to an existing directory tree.  When extracting to an empty directory tree, the new file will not exist.  So it will succeed always.

Also, since Android does not use hard links, this only matters for user created hard links.  In my case, I was extracting a debian chroot on /data/media without first deleting it.

Patch-set: 1