Update patch set 1
Patch Set 1: -Code-Review

I followed the code, I get your point

However, I couldn't follow to see if this exception is possible:

link(linktgt, filename) where linktgt doesn't exist?

I doubt there is any guaranty in the tar extract process to extract linktgt before trying to link filename

On the other side, extracting symlinks is safe as symlink() won't verify the path exists. It only writes the path string

What about adding an extra line to your actual patch:

https://github.com/omnirom/android_bootable_recovery/blob/android-4.4/libtar/extract.c#L365

However, you should add the same ENOENT check as you did for unlink

It will try to unlink and error if it couldn't on an existing file. If unlink is success, if the link() command fails because of non existing linktgt, we continue, else we abort

Hope I am not missing something else here as it is time to bed...

Patch-set: 1