Update patch set 1
Patch Set 1:

amend my above comment:

if the link() command fails because of non existing file, we should also abort I guess!

There is no way to fix this exception with hardlinks unless the --hard-dereference option was used, which is not a good default behavior anyway

continuing instead of aborting is a bad idea too
Another workaround is to modify the code to save the exception file missed. Later when extraction is done, it can try again to extract selected files

However, this should be done by caller to libtar source and not inside the libtar source

Patch-set: 1