Update patch set 1
Patch Set 1: Verified+1 Code-Review+2

The behavior here matches tar_extract_symlink().  Both have the potential issue that the target file is removed prior to calling (link/symlink).  So that if the (link/symlink) operation fails for any reason other than EEXIST, the target file is lost.  If you feel strongly that this is an issue that should be fixed, we should fix both at the same time.

Patch-set: 1