Update patch set 2
Patch Set 2:

Correct.  But note symlink() may also fail with ENOENT or any number of other cases, so the behavior of tar_extract_hardlink and tar_extract_symlink are equivalent now.

Again, if you feel this is worth fixing, you can change the code in both functions to check for EEXIST and do an unlink then retry.

Patch-set: 2