Update patch set 2
Patch Set 2:

Phyto, this is your change and it is only used by your recovery patch.  I can merge it whenever you like.

Also feel free to move libtar -> minitar after this gets submitted (or before, doesn't matter).  I didn't rename the directory for this change because it's easier to actually see what changed when the rename is a separate step.

Patch-set: 2