Update patch set 2
Patch Set 2:

This is a general purpose library that is used by at least two projects now. We should ensure that any error messages are shown on stderr if they are shown at all, regardless of how existing users behave after the function returns an error. Given that this is a library, it's not important to me that error messages are printed at all. That is the user's job. But I will not object if the messages go to stderr.

Yes I missed stdio.h in block.c. Sorry about that.

Patch-set: 2