Update patch set 3
Patch Set 3:

Fixed th_debug().

I think all of the perror calls should be in response to an actual error.  We should decide whether or not to show these errors when DEBUG is not defined.  Doesn't matter to me, so long as we're consistent in all the calls.

Of course, if we decide to show them only when DEBUG is defined, I'll write a wrapper.  Sprinkling "ifdef DEBUG" throughout the code is ugly.

Patch-set: 3