Update patch set 1
Patch Set 1:

> Wouldn't you need to define _FILE_OFFSET_BITS=64 to actually
 > support large files in libtar?

Only if the code needed to use a lib function that takes an off_t (such as lseek() or truncate()) without resorting to using the *64() variants.  But it doesn't need to use any of these lib functions, it only uses lstat() and, indirectly, read()/write().  And stat.st_size is 64-bit on all android platforms.

Patch-set: 1